HCRProbeMakerCommandLinev2.0
Major update!

There are no longer separate Mac/Linux and PC versions, these have been combined into a single version.

Whether trying to make probe sets for a single sequence or a full transcriptome, inputs are all done with fasta files.

Batch.py has now been incorporated into HCR.py and will intuit that you are seeking a batch based on the number of sequences in the input fasta.

BLAST has been reconfigured and should now work on Linux/Unix.

The code has been updated to account for dependency function deprecations.

A bug preventing the use of a single amplifier when batching has been squashed.

Making the same probe set with multiple amplifiers can be accomplished with the ampswap.py program, or by replicating the  sequence within the input fasta.

For batching, multiple amplifiers can be specified in a comma separated list (i.e. B1, b2, b4, B1).

Soft minimums on the number of probe pairs per set can be set, this simply discards input sequences that are shorter than what is necessary to meet the number input. This minimum is soft because if other parameters (polyAT or GC percentage for instance) cause a probe pair to be discarded, it is possible to end up with fewer pairs than the minimum requested.

There are surely other various bugs squashed and edits made, but these are the big changes.